validaciones controllers

+++ proyecto/controllers/productController.js
@@ -4,7 +4,10 @@ const product = db.products;
 const productController = {
 
     index: function (req, res) {
-        product.findAll()
+        product.findAll({ 
+            include: { all: true, nested: false }, 
+            order: [ ['id', 'DESC']],
+        })
             .then(function(products){
                 res.render('product_index', {products});
             })
@@ -26,17 +29,18 @@ const productController = {
     },
 
     add: function (req, res) {
+        if (!req.session.user) { 
+            throw Error('Not authorized.')
+        }
         res.render('add_product');
     },
 
     store: function(req, res) {
+        if (!req.session.user) { 
+            return res.render('add_product', { error: 'Not authorized.' });
+        }
         req.body.user_id = req.session.user.id;
-        
         if (req.file) req.body.img = (req.file.path).replace('public', '');
-        //image es el nombre del campo del formulario que carga la imagen
-        //para que venga el req.file primero le pusimos al formulario el enctype="multipart/form-data" 
-        //en req.file hay muchas propiedades y la mas importante es el path que tiene la ruta completa en donde esta el archivo
-        //el path lo aclaramos nosotoros en la carpeta destin   ation en la ruta
         product.create(req.body)
             .then(function() {
                 res.redirect('/')
@@ -44,9 +48,17 @@ const productController = {
             .catch(function(error) {
                 res.send(error);
             })
+        //image es el nombre del campo del formulario que carga la imagen
+        //para que venga el req.file primero le pusimos al formulario el enctype="multipart/form-data" 
+        //en req.file hay muchas propiedades y la mas importante es el path que tiene la ruta completa en donde esta el archivo
+        //el path lo aclaramos nosotoros en la carpeta destin   ation en la ruta
+        
     },
 
     delete: function(req, res) {
+        if (!req.session.user) { 
+            throw Error('Not authorized.')
+        } //chequear
         product.destroy({ where: { id: req.params.id } })
             .then(function() {
                 res.redirect('/')
